second commit

+++ deploy.sh
@@ -24,21 +24,27 @@ sudo swapon /swapfile
 # Make swap permanent
 echo '/swapfile none swap sw 0 0' | sudo tee -a /etc/fstab
 
-# Install Docker
-sudo apt install apt-transport-https ca-certificates curl software-properties-common -y
-curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
-sudo add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable" -y
-sudo apt update
-sudo apt install docker-ce -y
+# Check if docker installed
+docker-compose --version
+if [ $? -ne 0 ]; then
+  # Install Docker
+  sudo apt install apt-transport-https ca-certificates curl software-properties-common -y
+  curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
+  sudo add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable" -y
+  sudo apt update
+  sudo apt install docker-ce -y
 
-# Install Docker Compose
-sudo rm -f /usr/local/bin/docker-compose
-sudo curl -L "https://github.com/docker/compose/releases/download/v2.24.0/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
+  # Install Docker Compose
+  sudo rm -f /usr/local/bin/docker-compose
+  sudo curl -L "https://github.com/docker/compose/releases/download/v2.24.0/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose
 
+fi
 # Wait for the file to be fully downloaded before proceeding
 if [ ! -f /usr/local/bin/docker-compose ]; then
   echo "Docker Compose download failed. Exiting."
   exit 1
+
+
 fi
 
 sudo chmod +x /usr/local/bin/docker-compose
@@ -70,8 +76,8 @@ fi
 
 SQLALCHEMY_DATABASE_URI=postgresql+psycopg2://$POSTGRES_USER:$POSTGRES_PASSWORD@172.17.0.1:5432/$POSTGRES_DB
 # Create the .env file inside the app directory (~/myapp/.env)
-echo "SQLALCHEMY_DATABASE_URI=\"$SQLALCHEMY_DATABASE_URI\"" > "$SUPERSET_CONFIG_PATH"
-echo "SECRET_KEY=\"$SECRET_KEY" >> "$SUPERSET_CONFIG_PATH\""
+echo "SQLALCHEMY_DATABASE_URI="$SQLALCHEMY_DATABASE_URI"" > "$SUPERSET_CONFIG_PATH"
+echo "SECRET_KEY="$SECRET_KEY" >> "$SUPERSET_CONFIG_PATH""
 
 
 # Build and run the Docker containers from the app directory (~/myapp)
